+++ packages/bundler-webpack/src/config/handleNode.ts
@@ -1,11 +1,11 @@
-import type Config from 'webpack-5-chain'
+import type { Config } from 'webpack-v5-chain'
 
 /**
  * Set webpack node config
  */
 export const handleNode = ({ config }: { config: Config }): void => {
   // do not polyfill or mock node globals and modules
-  config.node.merge({
+  config.node({
     __filename: false,
     __dirname: false,
     global: false,