+++ packages/bundler-webpack/src/config/handleResolve.ts
@@ -1,5 +1,5 @@
 import type { App } from '@vuepress/core'
-import type Config from 'webpack-5-chain'
+import type { Config } from 'webpack-v5-chain'
 
 /**
  * Set webpack resolve
@@ -14,21 +14,26 @@ export const handleResolve = async ({
   isServer: boolean
 }): Promise<void> => {
   // aliases
-  config.resolve.alias
-    .set('@source', app.dir.source())
-    .set('@temp', app.dir.temp())
-    .set('@internal', app.dir.temp('internal'))
+  const alias = {
+    '@source': app.dir.source(),
+    '@temp': app.dir.temp(),
+    '@internal': app.dir.temp('internal'),
+  }
 
-  // extensions
-  config.merge({
-    resolve: {
-      extensionAlias: {
-        '.js': ['.js', '.ts'],
-        '.mjs': ['.mjs', '.mts'],
-      },
-      extensions: ['.js', '.jsx', '.ts', '.tsx', '.vue', '.json'],
-    },
+  // plugin hook: alias
+  const aliasResult = await app.pluginApi.hooks.alias.process(app, isServer)
+
+  aliasResult.forEach((aliasObject) => {
+    Object.assign(alias, aliasObject)
   })
+
+  // set aliases
+  config.resolve.alias.merge(
+    Object.fromEntries(
+      // sort alias by length in descending order to ensure longer alias is handled first
+      Object.entries(alias).sort(([a], [b]) => b.length - a.length),
+    ),
+  )
 
   // extensions
   config.resolve.extensions.merge([
@@ -40,13 +45,9 @@ export const handleResolve = async ({
     '.json',
   ])
 
-  // plugin hook: alias
-  const aliasResult = await app.pluginApi.hooks.alias.process(app, isServer)
-
-  // set aliases
-  aliasResult.forEach((aliasObject) => {
-    Object.entries(aliasObject).forEach(([key, value]) => {
-      config.resolve.alias.set(key, value as string)
-    })
+  // extensionAlias
+  config.resolve.extensionAlias.merge({
+    '.js': ['.js', '.ts'],
+    '.mjs': ['.mjs', '.mts'],
   })
 }