+++ packages/core/src/app/resolveAppPages.ts
@@ -1,4 +1,4 @@
-import { debug, globby } from '@vuepress/utils'
+import { debug, tinyglobby } from '@vuepress/utils'
 import { createPage } from '../page/index.js'
 import type { App, Page } from '../types/index.js'
 
@@ -15,7 +15,7 @@ export const resolveAppPages = async (
   log('resolveAppPages start')
 
   // resolve page absolute file paths according to the page patterns
-  const pageFilePaths = await globby(app.options.pagePatterns, {
+  const pageFilePaths = await tinyglobby.glob(app.options.pagePatterns, {
     absolute: true,
     cwd: app.dir.source(),
   })