+++ packages/core/tests/pluginApi/normalizeAliasDefineHook.spec.ts
@@ -1,6 +1,6 @@
 import { path } from '@vuepress/utils'
 import { expect, it, vi } from 'vitest'
-import type { AliasDefineHook, Bundler } from '../../src/index.js'
+import type { AliasHook, Bundler, DefineHook } from '../../src/index.js'
 import { createBaseApp, normalizeAliasDefineHook } from '../../src/index.js'
 
 const app = createBaseApp({
@@ -9,13 +9,19 @@ const app = createBaseApp({
   bundler: {} as Bundler,
 })
 
+it('should wrap object with a function', async () => {
+  const rawHook: AliasHook['exposed'] = {
+    foo: 'bar',
+  }
+  const normalizedHook = normalizeAliasDefineHook(rawHook)
+  expect(await normalizedHook(app, true)).toEqual({ foo: 'bar' })
+})
+
 it('should keep function as is', async () => {
-  const rawHook: AliasDefineHook['exposed'] = vi.fn(
-    (_app, isServer: boolean) => ({
-      foo: 'bar',
-      isServer,
-    }),
-  )
+  const rawHook: DefineHook['exposed'] = vi.fn((_app, isServer: boolean) => ({
+    foo: 'bar',
+    isServer,
+  }))
   const normalizedHook = normalizeAliasDefineHook(rawHook)
   expect(await normalizedHook(app, true)).toEqual({
     foo: 'bar',
@@ -23,12 +29,4 @@ it('should keep function as is', async () => {
   })
   expect(rawHook).toHaveBeenCalledTimes(1)
   expect(rawHook).toHaveBeenCalledWith(app, true)
-})
-
-it('should wrap object with a function', async () => {
-  const rawHook: AliasDefineHook['exposed'] = {
-    foo: 'bar',
-  }
-  const normalizedHook = normalizeAliasDefineHook(rawHook)
-  expect(await normalizedHook(app, true)).toEqual({ foo: 'bar' })
 })