+++ taze.config.js
@@ -3,7 +3,6 @@ import { defineConfig } from 'taze'
 export default defineConfig({
   packageMode: {
     '@types/express': 'minor',
-    'chokidar': 'minor',
     'express': 'minor',
     'vite': 'patch',
   },